this is working on the main section of the site to post lost items
The actual cards themselves are fine, but we need to create a button that loads html objects ( the cards) onto the posts page on the click of the button. you can use ajax to load new properties onto the page, but I have yet to figure out how to do it.