Create discussionBoard.php
here is the code that I have written up for the discussion board page- it is most of the php/html that would be needed to output a table and have the users submit text into the database to be printed out. if you could look over the section at the end that I commented on, that would be great. thanks