login/logout/db
I've added some code for the login and authentification portion of the project. It is similar in function to the code that we worked on in class. The main code that I added in is in the login.php page- that page combines the html of what initially was our landing page, with the php that is needed to login the user and validate their request to get to the posts page.